requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.100.0
flatbuffers==24.3.25
gast==0.4.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.10.0
idna==3.10
keras==2.15.0
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
ml-dtypes==0.2.0
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
python-multipart==0.0.9
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
# tensorflow-macos==2.15.0
termcolor==2.4.0
typing_extensions==4.8.0
urllib3==2.2.3
uvicorn==0.30.6
Werkzeug==3.0.4
wrapt==1.14.1
